retry.py: final clean-up
So this works fine with rsync, but scp fails to show progress as it
checks that the process group matches for it's tty. This is something
I can't work out how to set as the pty's created fail with the setpgrp
ioctl as "inappropriate". This is something to come back to later.

Also:

  * moves os to plain import (was getting messy)
  * deleted a bunch of commented out bits