retry.py: add bisect support
This allows me to do something like:

   git bisect run retry.py -t --timeout 20 -- "${QEMU_RUN[@]}"

And use retry as a wrapper to ensure build and test works. The timeout
allows me to test for hanging things which is quite common in QEMU land.

This does contain a horrible hack though for re-builds. Due to the way
buffering is handled I have a repeat the build step to account for the
times it fails because make can't write more to the stdout buffer.

I'm not sure where this problem comes from as in theory the sub-tasks
has control of the tty and the routing of the data.