retry.py: workaround terminal/build interaction
So if we have to SIGKILL QEMU we end up with a broken terminal. This
somehow affects make which starts complaining about not being able to
write to stdout and the whole point of automated bisecting goes out of
the window.

If we call out to "reset" before the controlling tty is grabbed back by
python then it will force the whole thing into the background. This
requires user intervention to get everything going again but after that
things continue as normal.

I'm not sure exactly where the bug is but it could certainly be
somewhere in the tty core which is very much a "here be dragons" area of
code.

FWIW I was running a bisection of my MTTCG code with the command line:

    git bisect run retry.py -b -t --timeout 20 -- "${QEMU_RUN[@]}"

Where "${QEMU_RUN[@]}" expands to:

    command=['/home/alex/lsrc/qemu/qemu.git/arm-softmmu/qemu-system-arm',
      '-machine', 'type=virt', '-display', 'none', '-smp', '1', '-m',
      '4096', '-cpu', 'cortex-a15', '-serial', 'telnet:127.0.0.1:4444',
      '-monitor', 'stdio', '-netdev',
      'user,id=unet,hostfwd=tcp::2222-:22',
      '-device', 'virtio-net-device,netdev=unet', '-drive',
      'file=/home/alex/lsrc/qemu/images/jessie-arm32.qcow2,id=myblock,index=0,if=none',
      '-device', 'virtio-blk-device,drive=myblock', '-append',
      'console=ttyAMA0 systemd.unit=benchmark.service root=/dev/vda1',
      '-kernel',
      '/home/alex/lsrc/qemu/images/aarch32-current-linux-kernel-only.img',
      '-smp', '2']

But I should probably come up with a better test case to reproduce the
bug.